+++ src/onegov.election_day/onegov/election_day/tests/views/test_views_election_compound.py
@@ -200,8 +200,9 @@ def test_view_election_compound_json(election_day_app_gr):
     login(client)
     upload_election_compound(client)
 
-    data = str(client.get('/elections/elections/json').json)
-    assert all((expected in data for expected in (
+    response = client.get('/elections/elections/json')
+    assert response.headers['Access-Control-Allow-Origin'] == '*'
+    assert all((expected in str(response.json) for expected in (
         "Carol", "Winner", "Hans", "Sieger"
     )))
 
@@ -215,7 +216,9 @@ def test_view_election_compound_summary(election_day_app_gr):
     with freeze_time("2014-01-01 12:00"):
         upload_election_compound(client)
 
-        assert client.get('/elections/elections/summary').json == {
+        response = client.get('/elections/elections/summary')
+        assert response.headers['Access-Control-Allow-Origin'] == '*'
+        assert response.json == {
             'completed': False,
             'date': '2015-01-01',
             'domain': 'canton',