+++ src/onegov.election_day/onegov/election_day/views/election/main.py
@@ -5,6 +5,7 @@ from onegov.core.security import Public
 from onegov.core.utils import normalize_for_url
 from onegov.election_day import ElectionDayApp
 from onegov.election_day.layouts import ElectionLayout
+from onegov.election_day.utils import add_cors_header
 from onegov.election_day.utils import add_last_modified_header
 from onegov.election_day.utils import get_election_summary
 from onegov.election_day.utils.election import get_candidates_results
@@ -35,7 +36,8 @@ def view_election_json(self, request):
     last_modified = self.last_modified
 
     @request.after
-    def add_last_modified(response):
+    def add_headers(response):
+        add_cors_header(response)
         add_last_modified_header(response, last_modified)
 
     embed = {}
@@ -181,7 +183,8 @@ def view_election_summary(self, request):
     """ View the summary of the election as JSON. """
 
     @request.after
-    def add_last_modified(response):
+    def add_headers(response):
+        add_cors_header(response)
         add_last_modified_header(response, self.last_modified)
 
     return get_election_summary(self, request)