+++ src/onegov.election_day/onegov/election_day/views/election_compound/main.py
@@ -5,6 +5,7 @@ from onegov.core.security import Public
 from onegov.core.utils import normalize_for_url
 from onegov.election_day import ElectionDayApp
 from onegov.election_day.layouts import ElectionCompoundLayout
+from onegov.election_day.utils import add_cors_header
 from onegov.election_day.utils import add_last_modified_header
 from onegov.election_day.utils import get_election_compound_summary
 
@@ -31,7 +32,8 @@ def view_election_compound_json(self, request):
     last_modified = self.last_modified
 
     @request.after
-    def add_last_modified(response):
+    def add_headers(response):
+        add_cors_header(response)
         add_last_modified_header(response, last_modified)
 
     embed = {}
@@ -87,7 +89,8 @@ def view_election_compound_summary(self, request):
     """ View the summary of the election compound as JSON. """
 
     @request.after
-    def add_last_modified(response):
+    def add_headers(response):
+        add_cors_header(response)
         add_last_modified_header(response, self.last_modified)
 
     return get_election_compound_summary(self, request)